Move the TaskDispatcher parts to a separate module
workflow_context.py is waaay too big as it is, but I can factor those
parts out to a separate module easily
This is just moving code around, not actually doing any changes to it.
Feel free to not read it too deeply, then.
